+++ tests/test_core.py
@@ -84,11 +84,11 @@ def no_wasm(note=''):
   return decorated
 
 
-def no_wasm2js(f):
-  def decorated(self):
-    if self.is_wasm_backend() and not self.get_setting('WASM'):
-      self.skipTest('wasm2js not supported')
-    f(self)
+def no_wasm2js(note=''):
+  assert not callable(note)
+
+  def decorated(f):
+    return skip_if(f, 'is_wasm2js', note)
   return decorated
 
 
@@ -154,6 +154,9 @@ def no_asan(note):
 
 
 class TestCoreBase(RunnerCore):
+  def is_wasm2js(self):
+    return self.is_wasm_backend() and not self.get_setting('WASM')
+
   # whether the test mode supports duplicate function elimination in js
   def supports_js_dfe(self):
     # wasm does this when optimizing anyhow
@@ -1928,6 +1931,7 @@ int main(int argc, char **argv) {
     self.do_run_in_out_file_test('tests', 'core', 'test_memorygrowth_3')
 
   @no_asmjs()
+  @no_wasm2js('no WebAssembly.Memory()')
   @no_asan('ASan alters the memory size')
   def test_module_wasm_memory(self):
     self.emcc_args += ['--pre-js', path_from_root('tests', 'core', 'test_module_wasm_memory.js')]
@@ -1974,7 +1978,7 @@ int main(int argc, char **argv) {
   def test_cxx03_do_run(self):
     self.do_run_in_out_file_test('tests', 'core', 'test_cxx03_do_run')
 
-  @no_wasm2js # massive switches can break js engines
+  @no_wasm2js('massive switches can break js engines')
   @no_emterpreter
   def test_bigswitch(self):
     src = open(path_from_root('tests', 'bigswitch.cpp')).read()
@@ -1984,7 +1988,7 @@ int main(int argc, char **argv) {
 3060: what?
 ''', args=['34962', '26214', '35040', str(0xbf4)])
 
-  @no_wasm2js # massive switches can break js engines
+  @no_wasm2js('massive switches can break js engines')
   @no_emterpreter
   @is_slow_test
   def test_biggerswitch(self):
@@ -4465,6 +4469,7 @@ Have even and odd!
   def test_transtrcase(self):
     self.do_run_in_out_file_test('tests', 'core', 'test_transtrcase')
 
+  @no_wasm2js('very slow to compile')
   def test_printf(self):
     # needs to flush stdio streams
     self.set_setting('EXIT_RUNTIME', 1)
@@ -5661,12 +5666,6 @@ return malloc(size);
       num_relocs = relocs.count('\n')
       return relocs, num_relocs
 
-    assert 'asm1' in core_test_modes
-    if self.run_name == 'asm1':
-      print('verifing relocations')
-      relocs, num_relocs = count_relocations()
-      self.assertEqual(num_relocs, 0)
-
     # TODO: wrappers for wasm modules
     if not self.is_wasm():
       print('relocatable')
@@ -5674,10 +5673,6 @@ return malloc(size);
       self.set_setting('RELOCATABLE', 1)
       self.set_setting('EMULATED_FUNCTION_POINTERS', 1)
       test()
-      if self.run_name == 'asm1':
-        relocs, num_relocs = count_relocations()
-        print('num_relocs: %s' % num_relocs)
-        self.assertGreater(num_relocs, 0)
       self.set_setting('RELOCATABLE', 0)
       self.set_setting('EMULATED_FUNCTION_POINTERS', 0)
 
@@ -6435,7 +6430,7 @@ return malloc(size);
     # Sanity check that it works and the dead function is emitted
     self.do_run(src, '*1*', args=['x'])
     js = open('src.cpp.o.js').read()
-    if self.run_name in ['default', 'asm1', 'asm2g']:
+    if self.run_name in ['default', 'asm2g']:
       assert 'function _unused($' in js
     self.do_run(None, '*2*', no_build=True)
 
@@ -6598,7 +6593,7 @@ return malloc(size);
                 ('|1.266,1|',                 # asm.js, double <-> int
                  '|1.266,1413754136|')) # wasm, reinterpret the bits
 
-  @no_wasm2js # TODO: nicely printed names in wasm2js
+  @no_wasm2js('TODO: nicely printed names in wasm2js')
   def test_demangle_stacks(self):
     self.set_setting('DEMANGLE_SUPPORT', 1)
     self.set_setting('ASSERTIONS', 1)
@@ -6983,7 +6978,7 @@ err = err = function(){};
 
   ### Tests for tools
 
-  @no_wasm2js # TODO: source maps in wasm2js
+  @no_wasm2js('TODO: source maps in wasm2js')
   @no_emterpreter
   def test_source_map(self):
     if not jsrun.check_engine(NODE_JS):
@@ -7860,6 +7855,7 @@ extern "C" {
     self.emcc_args += ['-s', 'USE_OFFSET_CONVERTER']
     self.do_run(open(path_from_root('tests', 'core', 'test_return_address.cpp')).read(), 'passed')
 
+  @no_wasm2js('TODO: sanitizers in wasm2js')
   @no_fastcomp('ubsan not supported on fastcomp')
   @no_asan('-fsanitize-minimal-runtime cannot be used with ASan')
   def test_ubsan_minimal_too_many_errors(self):
@@ -7872,6 +7868,7 @@ extern "C" {
     self.do_run(open(path_from_root('tests', 'core', 'test_ubsan_minimal_too_many_errors.c')).read(),
                 expected_output='ubsan: add-overflow\n' * 20 + 'ubsan: too many errors\n')
 
+  @no_wasm2js('TODO: sanitizers in wasm2js')
   @no_fastcomp('ubsan not supported on fastcomp')
   @no_asan('-fsanitize-minimal-runtime cannot be used with ASan')
   def test_ubsan_minimal_errors_same_place(self):
@@ -7890,6 +7887,7 @@ extern "C" {
     'fsanitize_overflow': (['-fsanitize=signed-integer-overflow'],),
   })
   @no_fastcomp('ubsan not supported on fastcomp')
+  @no_wasm2js('TODO: sanitizers in wasm2js')
   def test_ubsan_full_overflow(self, args):
     self.emcc_args += args
     self.do_run(open(path_from_root('tests', 'core', 'test_ubsan_full_overflow.c')).read(),
@@ -7902,6 +7900,7 @@ extern "C" {
     'fsanitize_undefined': (['-fsanitize=undefined'],),
     'fsanitize_return': (['-fsanitize=return'],),
   })
+  @no_wasm2js('TODO: sanitizers in wasm2js')
   @no_fastcomp('ubsan not supported on fastcomp')
   def test_ubsan_full_no_return(self, args):
     self.emcc_args += ['-Wno-return-type'] + args
@@ -7914,6 +7913,7 @@ extern "C" {
     'fsanitize_shift': (['-fsanitize=shift'],),
   })
   @no_fastcomp('ubsan not supported on fastcomp')
+  @no_wasm2js('TODO: sanitizers in wasm2js')
   def test_ubsan_full_left_shift(self, args):
     self.emcc_args += args
     self.do_run(open(path_from_root('tests', 'core', 'test_ubsan_full_left_shift.c')).read(),
@@ -7927,6 +7927,7 @@ extern "C" {
     'fsanitize_null': (['-fsanitize=null'],),
   })
   @no_fastcomp('ubsan not supported on fastcomp')
+  @no_wasm2js('TODO: sanitizers in wasm2js')
   def test_ubsan_full_null_ref(self, args):
     self.emcc_args += ['-std=c++11'] + args
     self.do_run(open(path_from_root('tests', 'core', 'test_ubsan_full_null_ref.cpp')).read(),
@@ -7941,6 +7942,7 @@ extern "C" {
     'fsanitize_vptr': (['-fsanitize=vptr'],),
   })
   @no_fastcomp('ubsan not supported on fastcomp')
+  @no_wasm2js('TODO: sanitizers in wasm2js')
   def test_ubsan_full_static_cast(self, args):
     self.emcc_args += args
     self.do_run(open(path_from_root('tests', 'core', 'test_ubsan_full_static_cast.cpp')).read(),
@@ -7961,6 +7963,7 @@ extern "C" {
     ]),
   })
   @no_fastcomp('ubsan not supported on fastcomp')
+  @no_wasm2js('TODO: sanitizers in wasm2js')
   def test_ubsan_full_stack_trace(self, g_flag, expected_output):
     self.emcc_args += ['-std=c++11', '-fsanitize=null', g_flag, '-s', 'ALLOW_MEMORY_GROWTH=1']
 
@@ -8060,6 +8063,7 @@ extern "C" {
                 expected_output=expected_output, assert_all=True,
                 check_for_error=False)
 
+  @no_wasm2js('TODO: ASAN in wasm2js')
   @no_fastcomp('asan not supported on fastcomp')
   def test_asan_js_stack_op(self):
     self.emcc_args += ['-fsanitize=address', '-s', 'ALLOW_MEMORY_GROWTH=1']
@@ -8153,7 +8157,6 @@ def make_run(name, emcc_args, settings=None, env=None):
 
 # Main asm.js test modes
 asm0 = make_run('asm0', emcc_args=[], settings={'ASM_JS': 2, 'WASM': 0})
-asm1 = make_run('asm1', emcc_args=['-O1'], settings={'WASM': 0})
 asm2 = make_run('asm2', emcc_args=['-O2'], settings={'WASM': 0})
 asm3 = make_run('asm3', emcc_args=['-O3'], settings={'WASM': 0})
 asm2g = make_run('asm2g', emcc_args=['-O2', '-g'], settings={'WASM': 0, 'ASSERTIONS': 1, 'SAFE_HEAP': 1})
@@ -8198,7 +8201,6 @@ wasm2s = make_run('wasm2s', emcc_args=['-O2'], settings={'SAFE_HEAP': 1})
 wasm2ss = make_run('wasm2ss', emcc_args=['-O2'], settings={'SAFE_STACK': 1})
 
 # emterpreter
-asmi = make_run('asmi', emcc_args=[], settings={'ASM_JS': 2, 'EMTERPRETIFY': 1, 'WASM': 0})
 asm2i = make_run('asm2i', emcc_args=['-O2'], settings={'EMTERPRETIFY': 1, 'WASM': 0})
 
 lsan = make_run('lsan', emcc_args=['-fsanitize=leak'], settings={'ALLOW_MEMORY_GROWTH': 1})